requirements.txt
discord.py[voice]==2.3.1
python-dotenv==1.0.0
mongoengine==0.27.0
pymongo==4.4.1
dnspython==2.4.1
pinecone-client==2.2.2
langchain==0.0.250
tiktoken==0.4.0
openai==0.27.8
tabulate==0.9.0
pypdf==3.14.0
playwright==1.36.0
unstructured==0.9.0
